src/requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
arabic-reshaper==3.0.0
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.17.0
fonttools==4.55.8
fsspec==2025.2.0
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_resources==6.5.2
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
mpmath==1.3.0
narwhals==1.24.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numpy==1.23.5
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
PyArabic==0.6.15
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyMuPDF==1.25.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.4
scipy==1.15.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
spaces==0.22.0
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.3
thop==0.1.1.post2209072238
tinycss2==1.4.0
tomlkit==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.1.8
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==11.0.3
yarg==0.1.9
